dfs yeah

+++ src/problem20.cpp
@@ -28,9 +28,9 @@ int dy[] = {0,0,1,-1};
 int main(){
     int N;
     cin >> N;
-    vector<int> A(N);
-    vector<int> B(N);
-    vector<int> C(N);
+    vector<ll> A(N);
+    vector<ll> B(N);
+    vector<ll> C(N);
     rep(i,N) cin >> A[i]; sort(A.begin(),A.end());
     rep(i,N) cin >> B[i]; 
     rep(i,N) cin >> C[i]; sort(C.begin(),C.end());
@@ -40,8 +40,8 @@ int main(){
     
     rep(i,N) {
 	
-	int a = lower_bound(A.begin(),A.end(),B[i]) - A.begin();
-	int c = N - (upper_bound(C.begin(),C.end(),B[i]) - C.begin());
+	ll a = lower_bound(A.begin(),A.end(),B[i]) - A.begin();
+	ll c = N - (upper_bound(C.begin(),C.end(),B[i]) - C.begin());
 	ans += a*c;
 //	cout << a << " " << c << endl;	
     }
